requirements.txt
RPi.GPIO
pyserial
requests
GitPython
